+++ declavatar/src/decl_v2/data/attachment.rs
