+++ declavatar/src/decl_v2/sexpr/da3/attachment.rs
@@ -1,6 +1,6 @@
 use crate::decl_v2::{
     data::{
-        arbittach::{DeclAttachment, DeclAttachmentProperty, DeclAttachmentValue, DeclAttachments},
+        attachment::{DeclAttachment, DeclAttachmentProperty, DeclAttachmentValue, DeclAttachments},
         StaticTypeName,
     },
     sexpr::{argument::SeparateArguments, error::KetosResult, register_function, KetosValueExt},
